+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 # 1.10.8
 
 * Ubuntu 18 and gcc7 are no longer supported.
+* Converted the decode mode from an arbitrary integer into a ProtoBuf enum.
+  This obviously breaks compatibility with older GTIRB files.
+
 
 # 1.10.7